renamed quick reply modal to quick post modal
