+++ src/components/quickPostModal/quickPostModal.styles.ts
